pyproject.toml
[project]
name = "gamecov"
version = "0.1.3"
description = "Coverage monitoring for directed game play-testing (Game-Fuzz)"
readme = "README.md"
authors = [{ name = "Yifeng He", email = "yfhe.prsn@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "av>=15.0.0",
    "fire>=0.7.0",
    "hypothesis>=6.136.7",
    "imagehash>=4.3.2",
    "imageio[ffmpeg]>=2.37.0",
    "numba>=0.61.2",
    "numpy>=2.0.0",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "pytest>=8.4.1",
    "returns[compatible-mypy]>=0.25.0",
    "scikit-image>=0.25.2",
    "stitching==0.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pylint.messages_control]
disable = [
    "raw-checker-failed",
    "bad-inline-option",
    "locally-disabled",
    "file-ignored",
    "suppressed-message",
    "useless-suppression",
    "deprecated-pragma",
    "use-symbolic-message-instead",
    "logging-fstring-interpolation",
    "unused-import",
    "ungrouped-imports",
    "subprocess-run-check",
    "too-many-instance-attributes",
    "missing-module-docstring",
]
fail-under = 9
[tool.pylint.typecheck]
generated-members = ["cv2.*", "skimage.metrics.*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
allow_redefinition = false
ignore_errors = false
implicit_reexport = false
local_partial_types = true
no_implicit_optional = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = false
plugins = ["returns.contrib.mypy.returns_plugin"]

# [tool.pytest.ini_options]
# pytest_plugins = ['pytest_profiling']

[dependency-groups]
dev = ["pytest-cov>=6.2.1", "pytest-profiling>=1.8.1"]
